+++ src/components/Cover.jsx
@@ -2,9 +2,42 @@
 // Licensed under the MIT License
 
 import React from 'react';
+import { getImage, withArtDirection } from 'gatsby-plugin-image';
 import styled from 'styled-components';
 
-export default (props) => {
+import SchemedImg from './SchemedImg';
+
+import theme from '../theme';
+
+const Cover = (props) => {
+  const { data } = props;
+
+  const imagesTopLight = withArtDirection(getImage(data.topPhoneSmallLight), [
+    {
+      media: theme.devices.tablet,
+      image: getImage(data.topTabletLight),
+    },
+  ]);
+  const imagesTopDark = withArtDirection(getImage(data.topPhoneSmallDark), [
+    {
+      media: theme.devices.tablet,
+      image: getImage(data.topTabletDark),
+    },
+  ]);
+
+  const imagesBottomLight = withArtDirection(getImage(data.bottomPhoneSmallLight), [
+    {
+      media: theme.devices.tablet,
+      image: getImage(data.bottomTabletLight),
+    },
+  ]);
+  const imagesBottomDark = withArtDirection(getImage(data.bottomPhoneSmallDark), [
+    {
+      media: theme.devices.tablet,
+      image: getImage(data.bottomTabletDark),
+    },
+  ]);
+
   const Section = styled.section`
     overflow: hidden;
     position: relative;
@@ -40,19 +73,42 @@ export default (props) => {
     text-align: center;
   `;
 
-  const ImgTop = () => props.imgTop;
-  const ImgBottom = () => props.imgBottom;
+  const Image = styled(SchemedImg)`
+    @media screen and ${(props) => props.theme.devices.tablet} {
+      width: ${(props) => props.tablet.width};
+      height: ${(props) => props.tablet.height};
+    }
+  `;
+
+  const imageStyles = {
+    position: 'absolute',
+    zIndex: -1,
+  };
 
   return (
     <Section>
-      <ImgTop />
+      <Image
+        light={imagesTopLight}
+        dark={imagesTopDark}
+        style={{ ...props.style.top, ...imageStyles }}
+        objectFit="cover"
+        tablet={{...props.tablet}}
+      />
       <Inner>
         <header>
           <Title>{props.title}</Title>
         </header>
         <Content>{props.children}</Content>
       </Inner>
-      <ImgBottom />
+      <Image
+        light={imagesBottomLight}
+        dark={imagesBottomDark}
+        style={{ ...props.style.bottom, ...imageStyles }}
+        objectFit="cover"
+        tablet={{...props.tablet}}
+      />
     </Section>
   );
 };
+
+export default Cover;