+++ src/components/StoryCover.jsx
@@ -3,13 +3,8 @@
 
 import React from 'react';
 import { graphql, useStaticQuery } from 'gatsby';
-import { getImage, withArtDirection } from 'gatsby-plugin-image';
-import styled from 'styled-components';
 
 import Cover from './Cover';
-import SchemedImg from './SchemedImg';
-
-import theme from '../theme';
 
 export default (props) => {
   const data = useStaticQuery(
@@ -19,130 +14,64 @@ export default (props) => {
           relativePath: { eq: "front-page/story/phone-small-up-right-light.png" }
         ) {
           childImageSharp {
-            gatsbyImageData(width: 300, height: 90, layout: FIXED)
+            gatsbyImageData(width: 300)
           }
         }
         topPhoneSmallDark: file(
           relativePath: { eq: "front-page/story/phone-small-up-right-dark.png" }
         ) {
           childImageSharp {
-            gatsbyImageData(width: 300, height: 90, layout: FIXED)
+            gatsbyImageData(width: 300)
           }
         }
         topTabletLight: file(relativePath: { eq: "front-page/story/tablet-up-right-light.png" }) {
           childImageSharp {
-            gatsbyImageData(width: 400, height: 400, layout: FIXED)
+            gatsbyImageData(width: 400)
           }
         }
         topTabletDark: file(relativePath: { eq: "front-page/story/tablet-up-right-dark.png" }) {
           childImageSharp {
-            gatsbyImageData(width: 400, height: 400, layout: FIXED)
+            gatsbyImageData(width: 400)
           }
         }
         bottomPhoneSmallLight: file(
           relativePath: { eq: "front-page/story/phone-small-down-left-light.png" }
         ) {
           childImageSharp {
-            gatsbyImageData(width: 300, height: 90, layout: FIXED)
+            gatsbyImageData(width: 300)
           }
         }
         bottomPhoneSmallDark: file(
           relativePath: { eq: "front-page/story/phone-small-down-left-dark.png" }
         ) {
           childImageSharp {
-            gatsbyImageData(width: 300, height: 90, layout: FIXED)
+            gatsbyImageData(width: 300)
           }
         }
         bottomTabletLight: file(
           relativePath: { eq: "front-page/story/tablet-down-left-light.png" }
         ) {
           childImageSharp {
-            gatsbyImageData(width: 400, height: 400, layout: FIXED)
+            gatsbyImageData(width: 400)
           }
         }
-        bottomTabletDark: file(relativePath: { eq: "front-page/story/tablet-down-left-dark.png" }) {
+        bottomTabletDark: file(
+          relativePath: { eq: "front-page/story/tablet-down-left-dark.png" }
+        ) {
           childImageSharp {
-            gatsbyImageData(width: 400, height: 400, layout: FIXED)
+            gatsbyImageData(width: 400)
           }
         }
       }
     `,
   );
 
-  const imagesTopLight = withArtDirection(getImage(data.topPhoneSmallLight), [
-    {
-      media: theme.devices.tablet,
-      image: getImage(data.topTabletLight),
-    },
-  ]);
-  const imagesTopDark = withArtDirection(getImage(data.topPhoneSmallDark), [
-    {
-      media: theme.devices.tablet,
-      image: getImage(data.topTabletDark),
-    },
-  ]);
-
-  const imagesBottomLight = withArtDirection(getImage(data.bottomPhoneSmallLight), [
-    {
-      media: theme.devices.tablet,
-      image: getImage(data.bottomTabletLight),
-    },
-  ]);
-  const imagesBottomDark = withArtDirection(getImage(data.bottomPhoneSmallDark), [
-    {
-      media: theme.devices.tablet,
-      image: getImage(data.bottomTabletDark),
-    },
-  ]);
-
-  const Img = styled(SchemedImg)`
-    margin: 0;
-
-    @media screen and ${(props) => props.theme.devices.tablet} {
-      margin: 0 4em;
-    }
-  `;
-
-  const TopImg = styled(Img)`
-    @media screen and ${(props) => props.theme.devices.tablet} {
-      width: 400px;
-      height: 400px;
-    }
-  `;
-
-  const BottomImg = styled(Img)`
-    @media screen and ${(props) => props.theme.devices.tablet} {
-      width: 400px;
-      height: 400px;
-    }
-  `;
-
-  const imgStyles = {
-    position: 'absolute',
-    zIndex: -1,
-  };
-
   return (
     <Cover
-      title={props.title}
-      imgTop={
-        <TopImg
-          imageLight={imagesTopLight}
-          imageDark={imagesTopDark}
-          style={{ top: 0, right: 0, ...imgStyles }}
-          objectFit="cover"
-        />
-      }
-      imgBottom={
-        <BottomImg
-          imageLight={imagesBottomLight}
-          imageDark={imagesBottomDark}
-          style={{ bottom: 0, left: 0, ...imgStyles }}
-          objectFit="cover"
-        />
-      }
-    >
-      {props.children}
-    </Cover>
+      data={data}
+      tablet={{ width: '400px', height: '400px' }}
+      style={{ top: {top: 0, right: 0}, bottom: { bottom: 0, left: 0 } }}
+      {...props}
+    />
   );
 };